app/requirements.txt
boto3==1.42.41
botocore==1.42.41
urllib3==2.6.3
requests==2.31.0
python-dotenv==1.0.1
aws-secretsmanager-caching==1.1.1.5
pydantic
pydantic[email]
pytest
pytz==2025.2
google-auth==2.29.0
google-auth-httplib2==0.2.0
